web-animations/animation-sequence/animation-sequence-children-start-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationSequence child nodes start time test</title>
<meta name="assert" content="The start time of a child animation node of an animation sequence is the end time of the child's previous sibling. If the child has no previous sibling the start time is zero">
<link rel="help" href="dev.w3.org/fxtf/web-animations/#animation-sequences">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
async_test(function(t) {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationSequence = new AnimationSequence([node1, node2]);
    var player = document.timeline.play(animationSequence);

    player.ready.then(t.step_func(function() {
        assert_equals(node1.computedTiming.startTime, 0, 'The start time of the first ' +
            'child animation node of an animation sequence should be zero');
        assert_equals(node2.computedTiming.startTime, node1.computedTiming.endTime,
            'The start time of the second child animation node of an animation sequence ' +
            'should be the end time of its previous sibling');
        t.done();
    }));
}, 'The start time of a child animation node of an animation sequence is the end time ' +
    'of the child\'s previous sibling. If the child has no previous sibling the start ' +
    'time is zero');

async_test(function(t) {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = newAnimation(createDiv(this));
    var node5 = new AnimationGroup([node1, node2]);
    var node6 = new AnimationSequence([node3]);
    var animationSequence = new AnimationSequence([node4, node5, node6]);
    var player = document.timeline.play(animationSequence);

    player.ready.then(t.step_func(function() {
       assert_equals(node4.computedTiming.startTime, 0, 'The start time of the ' +
           'first child of an animation group should be zero');
       assert_equals(node5.computedTiming.startTime, node4.computedTiming.endTime,
           'The start time of an animation node should be the end time of its previous sibling');
       assert_equals(node6.computedTiming.startTime, node5.computedTiming.endTime,
           'The start time of an animation node should be the end time of its previous sibling');
       assert_equals(node1.computedTiming.startTime, 0, 'The start time of a child ' +
           'animation node of an animation group should be zero');
       assert_equals(node2.computedTiming.startTime, 0, 'The start time of a child ' +
           'animation node of an animation group should be zero');
       assert_equals(node3.computedTiming.startTime, 0, 'The start time of the ' +
           'first child of an animation group should be zero');
       t.done();
   }));
}, 'Test the start time of a child animation nodes of an animation sequence. ' +
    'Test different types of not empty child nodes');
</script>
</body>
